Task_Management_System/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
drf-yasg
python-dotenv
psycopg2-binary
django-cors-headers
